# emas/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.ReferenceLocation:
    properties:
      bay:
        type: string
      display:
        description: computed in app
        type: string
      id:
        type: integer
      zone:
        type: string
    type: object
  domain.ReferenceMachineType:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  domain.ReferenceProductType:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.CreateLocationRequest:
    properties:
      bay:
        type: string
      zone:
        type: string
    required:
    - zone
    type: object
...
  dto.CreateMachineTypeRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  dto.CreateProductTypeRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  dto.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  dto.UpdateMachineTypeRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Factory operations API.
  title: eMas Factory API
  version: "1.0"
paths:
...
  /reference/locations:
    get:
      consumes:
      - application/json
      description: List locations
      parameters:
      - description: Search by zone or bay
        in: query
        name: q
        type: string
      - description: Sort field (id, zone, bay)
        in: query
        name: sort_by
        type: string
      - description: Sort direction (asc, desc)
        in: query
        name: sort_dir
        type: string
      - description: Limit number of results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Comma-separated fields to return
        in: query
        name: fields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ReferenceLocation'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List locations
      tags:
      - reference
    post:
      consumes:
      - application/json
      description: Create a location
      parameters:
      - description: Create Location Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLocationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ReferenceLocation'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Create a location
      tags:
      - reference
  /reference/locations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a location
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Delete a location
      tags:
      - reference
  /reference/machine-types:
    get:
      consumes:
      - application/json
      description: List machine types
      parameters:
      - description: Search by name
        in: query
        name: q
        type: string
      - description: Sort field (id, name)
        in: query
        name: sort_by
        type: string
      - description: Sort direction (asc, desc)
        in: query
        name: sort_dir
        type: string
      - description: Limit number of results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Comma-separated fields to return
        in: query
        name: fields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ReferenceMachineType'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List machine types
      tags:
      - reference
    post:
      consumes:
      - application/json
      description: Create a machine type
      parameters:
      - description: Create Machine Type Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMachineTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ReferenceMachineType'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Create a machine type
      tags:
      - reference
  /reference/machine-types/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a machine type
      parameters:
      - description: Machine Type ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Delete a machine type
      tags:
      - reference
    put:
      consumes:
      - application/json
      description: Update a machine type
      parameters:
      - description: Machine Type ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Machine Type Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMachineTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ReferenceMachineType'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Update a machine type
      tags:
      - reference
  /reference/product-types:
    get:
      consumes:
      - application/json
      description: List product types
      parameters:
      - description: Search by name
        in: query
        name: q
        type: string
      - description: Sort field (id, name)
        in: query
        name: sort_by
        type: string
      - description: Sort direction (asc, desc)
        in: query
        name: sort_dir
        type: string
      - description: Limit number of results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Comma-separated fields to return
        in: query
        name: fields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ReferenceProductType'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List product types
      tags:
      - reference
    post:
      consumes:
      - application/json
      description: Create a product type
      parameters:
      - description: Create Product Type Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ReferenceProductType'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Create a product type
      tags:
      - reference
  /reference/product-types/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product type
      parameters:
      - description: Product Type ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Delete a product type
      tags:
      - reference
...
swagger: "2.0"